Adds more tests, improved naming conventions.
