--- ftest-toxic-frontend-reporter/pom.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>arquillian-cube-q</artifactId>
-    <groupId>org.arquillian.cube.q</groupId>
-    <version>1.0.0.Final-SNAPSHOT</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>arquillian-cube-q-recorder-ftest-toxic-frontend</artifactId>
-
-
-  <name>Arquillian Cube Q Toxic Reporter FrontEnd Functional Test</name>
-
-  <properties>
-    <maven.compiler.target>1.8</maven.compiler.target>
-    <maven.compiler.source>1.8</maven.compiler.source>
-  </properties>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.jboss.arquillian</groupId>
-        <artifactId>arquillian-bom</artifactId>
-        <version>${version.arquillian_core}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.arquillian.cube.q</groupId>
-      <artifactId>arquillian-cube-q-toxic</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.arquillian.junit</groupId>
-      <artifactId>arquillian-junit-standalone</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.arquillian.cube.q</groupId>
-      <artifactId>arquillian-cube-q-recorder</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.arquillian.extension</groupId>
-      <artifactId>arquillian-recorder-reporter-impl</artifactId>
-    </dependency>
-  </dependencies>
-
-</project>
