--- recorder/pom.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>arquillian-cube-q</artifactId>
-    <groupId>org.arquillian.cube.q</groupId>
-    <version>1.0.0.Final-SNAPSHOT</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>arquillian-cube-q-recorder</artifactId>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.arquillian.cube.q</groupId>
-      <artifactId>arquillian-cube-q-spi</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.arquillian.cube.q</groupId>
-      <artifactId>arquillian-cube-q-toxic</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.mockito</groupId>
-      <artifactId>mockito-core</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.assertj</groupId>
-      <artifactId>assertj-core</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-</project>
