+++ reporter/src/main/java/org/arquillian/cube/q/recorder/CubeQRecorderExtension.java
