+++ reporter/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
